lilylee_빌드 오류 체크하고, 배포가능 확인
