+++ middleware.off.ts
@@ -3,8 +3,8 @@
 import { clerkMiddleware } from "@clerk/nextjs/server";
 import { NextResponse } from "next/server";
 
-export default clerkMiddleware((auth, req) => {
-  const userId = auth().userId;  //clerk v6.x 에서 호출금지 오류가 있음. 
+export default clerkMiddleware(async (auth, req) => {
+  const { userId } = await auth();  //clerk v6.x 에서 호출금지 오류가 있음. 
   const { pathname } = req.nextUrl;
 
   // 루트 접속 분기
